requirements.txt
annotated-types==0.7.0
anyio==4.8.0
azure-core==1.32.0
azure-functions==1.21.3
azure-storage-blob==12.24.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
distro==1.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
isodate==0.7.2
jiter==0.8.2
openai==1.59.6